requirements.txt
fastapi
uvicorn[standard]
pandas
matplotlib
requests
beautifulsoup4
duckdb
pyarrow
pillow
openai
python-multipart